+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 
 - Broker graceful shutdown with task tracking
 - Broker fixes and updates
+- Signal and queue hash over name
 
 1.8.2 (2026-06-11)
 ==================